+++ common/src/main/kotlin/utils/Argon2.kt
@@ -5,10 +5,10 @@ import de.mkammerer.argon2.Argon2Helper
 import kotlin.time.Duration
 
 class Argon2(
-    val memory: Int = DEFAULT_MEMORY,
+    val memory: Long = DEFAULT_MEMORY,
     val iterations: Int,
 ) {
-    val memoryInKiB = memory / 1024
+    val memoryInKiB = (memory / 1024).toInt()
 
     fun hash(password: CharArray): String {
         return argon2.hash(iterations, memoryInKiB, CORES, password)
@@ -20,12 +20,12 @@ class Argon2(
 
     companion object {
         private val CORES = Runtime.getRuntime().availableProcessors()
-        val DEFAULT_MEMORY = 64.MiB.toInt()
+        val DEFAULT_MEMORY = 64.MiB
 
         private val argon2 = Argon2Factory.create(Argon2Factory.Argon2Types.ARGON2id)
 
-        fun findIterations(maxTime: Duration, memory: Int = DEFAULT_MEMORY): Int {
-            return Argon2Helper.findIterations(argon2, maxTime.inWholeMilliseconds, memory / 1024, CORES)
+        fun findIterations(maxTime: Duration, memory: Long = DEFAULT_MEMORY): Int {
+            return Argon2Helper.findIterations(argon2, maxTime.inWholeMilliseconds, (memory / 1024).toInt(), CORES)
         }
     }
 }
